+++ src/main/kotlin/io/github/siyual_park/domain/scope/ScopeResponsePayloadMapper.kt
@@ -15,7 +15,6 @@ class ScopeResponsePayloadMapper(
             name,
             description,
             scopeFetchExecutor.execute(this, 1)
-                .filter { it != this }
                 .map { child ->
                     ScopeTokenResponsePayload(
                         child.id!!,
