+++ src/main/resources/application.properties
@@ -8,4 +8,4 @@ spring.jpa.properties.hibernate.format_sql=true
 
 application.token.secret=5MSfAxtWOwgLPKcYr5Le6mKlv56cGK2gwZEAfWudaT5BAF_9cyS2lz3FxNjMpJ
 application.token.expiresIn=360
-application.scope.default=role:default
+application.scope.default=pack:default
